Add `hyper 0.8.1` compatibility
Hyper.app (F.K.A. HyperTerm) is now on version 0.8.1 and contains
modifications in markup and styling of the `nav`/`tabs_list` elements.
This causes a space to appear in place of the traffic lights location.
However the aberration on the `tab_first` element has disappeared and
makes the `hideFirstTabBorder` configuration option unnecessary.